requirements.txt
accelerate==0.10.0
albumentations==1.2.0
datasets==2.3.2
numpy==1.22.3
omegaconf==2.2.2
opencv_python_headless==4.6.0.66
optuna==2.10.1
Pillow==9.2.0
scikit_learn==1.1.1
torch==1.12.0
tqdm==4.64.0
transformers==4.18.0
typer==0.6.1
